Tests: Moved repeat operator tests to test_grammar
